Update package setup
