+++ tests/.meteor/packages
@@ -1,0 +1,30 @@
+# Meteor packages used by this project, one per line.
+# Check this file (and the other files in this directory) into your repository.
+#
+# 'meteor add' and 'meteor remove' will edit this file for you,
+# but you can also edit it by hand.
+
+meteor-base@1.5.2-beta300.0             # Packages every Meteor app needs to have
+mongo@2.0.0-beta300.0                   # The database Meteor supports right now
+reactive-var@1.0.13-beta300.0            # Reactive variable for tracker
+jquery                  # Helpful client-side library
+tracker@1.3.3-beta300.0                 # Meteor's client-side reactive programming library
+
+standard-minifier-css@1.9.3-beta300.0   # CSS minifier run for production mode
+standard-minifier-js@3.0.0-beta300.0    # JS minifier run for production mode
+es5-shim@4.8.1-beta300.0                # ECMAScript 5 compatibility for older browsers.
+ecmascript@0.16.8-beta300.0              # Enable ECMAScript2015+ syntax in app code
+shell-server@0.6.0-beta300.0            # Server-side component of the `meteor shell` command
+
+autopublish@1.0.8-beta300.0             # Publish all data to the clients (for prototyping)
+insecure@1.0.8-beta300.0                # Allow all DB writes from clients (for prototyping)
+
+underscore@1.0.14-beta300.0
+dynamic-import@0.7.4-beta300.0
+
+logging
+quave:synced-cron
+aldeed:simple-schema@1.13.1
+aldeed:collection2@4.0.0-beta.6
+meteortesting:mocha@3.1.0-beta300.0
+meteortesting:mocha-core@8.3.0-beta300.0
