+++ tests/.meteor/platforms
@@ -1,0 +1,2 @@
+server
+browser
