+++ tests/.meteor/release
@@ -1,0 +1,1 @@
+METEOR@3.0-beta.0
