+++ tests/.meteor/versions
@@ -1,0 +1,70 @@
+aldeed:collection2@4.0.0-beta.6
+aldeed:simple-schema@1.13.1
+allow-deny@2.0.0-beta300.0
+autopublish@1.0.8-beta300.0
+autoupdate@2.0.0-beta300.0
+babel-compiler@7.11.0-beta300.0
+babel-runtime@1.5.2-beta300.0
+base64@1.0.13-beta300.0
+binary-heap@1.0.12-beta300.0
+boilerplate-generator@2.0.0-beta300.0
+callback-hook@1.6.0-beta300.0
+check@1.3.3-beta300.0
+core-runtime@1.0.0-beta300.0
+ddp@1.4.2-beta300.0
+ddp-client@3.0.0-beta300.0
+ddp-common@1.4.1-beta300.0
+ddp-server@3.0.0-beta300.0
+diff-sequence@1.1.3-beta300.0
+dynamic-import@0.7.4-beta300.0
+ecmascript@0.16.8-beta300.0
+ecmascript-runtime@0.8.2-beta300.0
+ecmascript-runtime-client@0.12.2-beta300.0
+ecmascript-runtime-server@0.11.1-beta300.0
+ejson@1.1.4-beta300.0
+es5-shim@4.8.1-beta300.0
+facts-base@1.0.2-beta300.0
+fetch@0.1.4-beta300.0
+geojson-utils@1.0.12-beta300.0
+hot-code-push@1.0.5-beta300.0
+http@1.0.1
+id-map@1.2.0-beta300.0
+insecure@1.0.8-beta300.0
+inter-process-messaging@0.1.2-beta300.0
+jquery@3.0.0
+logging@1.3.3-beta300.0
+meteor@2.0.0-beta300.0
+meteor-base@1.5.2-beta300.0
+meteortesting:browser-tests@1.6.0-beta300.0
+meteortesting:mocha@2.1.0
+meteortesting:mocha-core@8.3.1-beta300.0
+minifier-css@2.0.0-beta300.0
+minifier-js@3.0.0-beta300.0
+minimongo@2.0.0-beta300.0
+modern-browsers@0.1.10-beta300.0
+modules@0.19.1-beta300.0
+modules-runtime@0.13.2-beta300.0
+mongo@2.0.0-beta300.0
+mongo-decimal@0.1.4-beta300.0
+mongo-dev-server@1.1.1-beta300.0
+mongo-id@1.0.9-beta300.0
+npm-mongo@4.16.1-beta300.0
+ordered-dict@1.2.0-beta300.0
+promise@1.0.0-beta300.0
+quave:synced-cron@2.0.2
+raix:eventemitter@1.0.0
+random@1.2.2-beta300.0
+react-fast-refresh@0.2.8-beta300.0
+reactive-var@1.0.13-beta300.0
+reload@1.3.2-beta300.0
+retry@1.1.1-beta300.0
+routepolicy@1.1.2-beta300.0
+shell-server@0.6.0-beta300.0
+socket-stream-client@0.5.2-beta300.0
+standard-minifier-css@1.9.3-beta300.0
+standard-minifier-js@3.0.0-beta300.0
+tracker@1.3.3-beta300.0
+typescript@4.9.5-beta300.0
+underscore@1.0.14-beta300.0
+webapp@2.0.0-beta300.0
+webapp-hashing@1.1.2-beta300.0
